backend-flask/requirements.txt
flask
flask-cors

opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-http

aws-xray-sdk

watchtower

blinker
rollbar